(PC-17493) navigation helpers: separate helpers.test in different files
	- isAppUrl.native.test.ts
	- navigateToBooking.native.test.ts
	- openUrl.native.test.ts
